Adapting KeyValueListControl for better imports
