+++ client/packages/lowcoder/src/comps/comps/allComp.test.tsx
@@ -51,16 +51,16 @@ const COMPS_MAP = {
   string: StringControl,
   test: TestComp,
   input: InputComp,
-  // query: QueryComp,
-  // js: JSQuery,
-  // mysql: SQLQuery,
-  // params: ParamsStringControl,
-  // layout: AppLayoutComp,
-  // root: RootComp,
-  // listView: ListViewComp,
-  // table: TableComp,
-  // tableColumn: ColumnComp,
-  // tableColumnRender: RenderComp,
+  query: QueryComp,
+  js: JSQuery,
+  mysql: SQLQuery,
+  params: ParamsStringControl,
+  layout: AppLayoutComp,
+  root: RootComp,
+  listView: ListViewComp,
+  table: TableComp,
+  tableColumn: ColumnComp,
+  tableColumnRender: RenderComp,
 } as Record<string, CompConstructor>;
 
 Object.entries(uiCompRegistry).forEach(async ([key, value]) => {
