+++ client/packages/lowcoder/src/comps/controls/keyValueControl.tsx
@@ -1,13 +1,11 @@
-import { ControlPropertyViewWrapper, KeyValueList, OptionsType } from "lowcoder-design";
+import { OptionsType } from "lowcoder-design";
 import { ReactNode } from "react";
 import styled from "styled-components";
 import { MultiCompBuilder } from "../generators";
-import { list } from "../generators/list";
 import { StringControl } from "./codeControl";
 import { ControlParams } from "./controlParams";
 import { dropdownControl } from "./dropdownControl";
-import { ParamsControlType, ParamsStringControl } from "./paramsControl";
-import { FunctionProperty } from "../queries/queryCompUtils";
+import { ParamsStringControl } from "./paramsControl";
 
 const KeyValueWrapper = styled.div`
   display: flex;
@@ -57,7 +55,7 @@ export type KeyValueControlParams = ControlParams & {
  * controlType: params output: {key: {"1+2": () => "3"}, value: {"-1": () => "-1"}}
  * controlType: string output: {key: "xxx", value: "xxxx"}
  */
-function keyValueControl<T extends OptionsType>(
+export function keyValueControl<T extends OptionsType>(
   hasType: boolean = false,
   types: T,
   controlType: "params" | "string" = "params"
@@ -112,42 +110,3 @@ function keyValueControl<T extends OptionsType>(
   };
 }
 
-/**
- * Provides a list of kv input boxes with add and delete buttons
- * output [{key: "", value: ""}, {key: "", value: ""}]
- */
-export function keyValueListControl<T extends OptionsType>(
-  hasType: boolean = false,
-  types: T | OptionsType = [],
-  controlType: "params" | "string" = "params"
-) {
-  return class extends list(keyValueControl(hasType, types, controlType)) {
-    getQueryParams() {
-      if (controlType === "params") {
-        return this.getView().reduce(
-          (result: FunctionProperty[], kv) => [
-            ...result,
-            ...(kv.children.key as InstanceType<ParamsControlType>).getQueryParams(),
-            ...(kv.children.value as InstanceType<ParamsControlType>).getQueryParams(),
-          ],
-          []
-        );
-      }
-      return [];
-    }
-
-    propertyView(params: KeyValueControlParams): ReactNode {
-      return (
-        <ControlPropertyViewWrapper {...params}>
-          <KeyValueList
-            list={this.getView().map((child) => child.propertyView(params))}
-            onAdd={() => this.dispatch(this.pushAction({}))}
-            onDelete={(item, index) => this.dispatch(this.deleteAction(index))}
-            isStatic={params.isStatic}
-            indicatorForAll={params.indicatorForAll}
-          />
-        </ControlPropertyViewWrapper>
-      );
-    }
-  };
-}
