+++ client/packages/lowcoder/src/i18n/locales/en.ts
@@ -855,7 +855,8 @@ export const en = {
     "categoryDocumentHandling" : "Report & Document Generation",
     "categoryRPA" : "Robotic Process Automation",
     "componentsUsingQueryTitle" : "Query Usage",
-    "componentsUsingQuery" : "Where is this Query in use"
+    "componentsUsingQuery" : "Where is this Query in use",
+    "query.variables" : "Variables"
   },
