+++ client/packages/lowcoder/src/index.sdk.ts
@@ -93,6 +93,7 @@ export * from "comps/controls/actionSelector/actionSelectorControl";
 export * from "comps/controls/iconControl";
 export * from "comps/controls/shapeControl";
 export * from "comps/controls/keyValueControl";
+export * from "comps/controls/keyValueListControl";
 export * from "comps/controls/labelControl";
 export * from "comps/controls/millisecondControl";
 export * from "comps/controls/optionsControl";
